requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==2.1.9
Django==3.2.19
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
